Added ValheimHelper class
